refactor(admin): AdminInstitutions sous 300 lignes — éclatement complet (#G1 ≤300)
AdminInstitutions.vue : 666 -> 205. Composable useAdminInstitutions (état + load +
create/edit/save/toggle/testConnection) ; sous-composants InstitutionsStatsCards et
InstitutionsTable avec leur CSS. La vue orchestre header + états + table + InstitutionModals
(déjà extraite). formatDate converge vers formatDateShort (util partagé, parité exacte
null→'Aucune' et dates valides). CSS distribué.

Tests : useAdminInstitutions (11 cas, API mockée) + InstitutionsStatsCards + InstitutionsTable.
21/21 verts (avec InstitutionModals non-régression). vite build vert. Fichiers ≤300.
